+++ R/validate_inputs.R
@@ -25,7 +25,7 @@ read_iso3 <- function(file, type) {
   func <- switch(type,
     "pjnz" = read_pjnz_iso3,
     "shape" = read_geojson_iso3,
-    stop(sprintf("Can't read country from data of type %s.", type)))
+    stop(t_("invalid_iso3", data = list(type = type))))
   func(file)
 }
 
@@ -36,7 +36,7 @@ read_regions <- function(file, type) {
     "programme" = read_csv_regions,
     "anc" = read_csv_regions,
     "survey" = read_csv_regions,
-    stop(sprintf("Can't read regions from data of type %s.", type)))
+    stop(t_("invalid_regions", data = list(type = type))))
   func(file)
 }