+++ R/zzz.R
@@ -4,18 +4,26 @@
 #' @importFrom jsonlite unbox
 #' @importFrom naomi get_metadata
 #' @importFrom rrq rrq_controller
+#' @importFrom traduire t_
 NULL
 
 
 cfg <- new.env(parent = emptyenv())
 .onLoad <- function(...) {
-  get_version_info() # nocov
+  cfg$version_info <- get_version_info() # nocov
+  hintr_init_traduire() # nocov
 }
 
 get_version_info <- function() {
-  packages <- c("hintr", "naomi", "rrq")
+  packages <- c("hintr", "naomi", "rrq", "traduire")
   value <- lapply(packages, function(p)
     scalar(as.character(utils::packageVersion(p))))
   names(value) <- packages
-  cfg$version_info <- value
+  value
+}
+
+hintr_init_traduire <- function() {
+  path <- system.file("traduire/translations.json",
+                      package = "hintr", mustWork = TRUE)
+  traduire::translator_register(path, "en")
 }